NIFI-14675 Support long number literals in RecordPath. Compare long numbers without precision loss in RecordPath filters.

+++ nifi-commons/nifi-record-path/src/main/java/org/apache/nifi/record/path/paths/RecordPathCompiler.java
@@ -219,7 +219,7 @@ public class RecordPathCompiler {
                 return new LiteralValuePath(parent, tree.getText(), absolute);
             }
             case NUMBER: {
-                return new LiteralValuePath(parent, Integer.parseInt(tree.getText()), absolute);
+                return new LiteralValuePath(parent, Long.parseLong(tree.getText()), absolute);
             }
             case PREDICATE: {
                 final Tree operatorTree = tree.getChild(0);
